build(android): rename application id to com.echo_messenger for play store
