+++ apps/client/android/app/src/main/kotlin/com/echo_messenger/EchoForegroundService.kt
@@ -1,4 +1,4 @@
-package us.echomessenger.echo_app
+package com.echo_messenger
 
 import android.app.Notification
 import android.app.NotificationChannel
