+++ apps/client/android/app/src/main/kotlin/com/echo_messenger/MainActivity.kt
@@ -1,4 +1,4 @@
-package us.echomessenger.echo_app
+package com.echo_messenger
 
 import android.app.PictureInPictureParams
 import android.content.Intent
